+++ CHANGELOG.md
@@ -637,6 +637,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 - Use 2 bit record the number of occurrences of 1: `00`, `01`, `10`
 - Use **Karnaugh map** to facilitate the simplification of Boolean algebra
 
+## 0138 - Copy List with Random Pointer
+
+- Use Map to store old node to new node
+
 ## 0141 - Linked List Cycle
 
 - Two points: slow and fast